Further updates for TM extraction.
There is still a gap, since we need to know the initial connection
TM frame as the first header pointer might be set. Not sure how to
deal with this. Somehow we need to pass information down the line
for the first TM frame, which should ignore the data before the
first header pointer. Maybe this can be done within the conduit itself.